Add support for a MediaWiki:GlobalUserPage.css
Many templates require CSS styles to display as intended, however
those styles are usually in MediaWiki:Common.css, meaning that
remote wikis may not have them. So, we add MediaWiki:GlobalUserPage.css,
which can contain those styles. It is loaded on the central
wiki's user page, as well as when displaying the remote page.

Change-Id: Ib31523bd973277ff3f8a86dbe96108a88cdc9975