Allow opt-out via <noinclude>
Instead of getting the parsed HTML for the user page itself,
now get it for "{{:User:$name}}", which handles <noinclude>
tags. If the rendered HTML is empty, we don't display a
global user page.

This will also allow people to use <includeonly> tags to
only display content on a global user page.

We also no longer include the newPP report in the displayed HTML.

Change-Id: I6bf0e6e84e589a47fd9f35da35db8b60edb85c41