Implement magic word based opt-out
The current method of opting out via <noinclude> tags is a bit hacky and
has a few problems:
* It requires transcluding the page itself
* Determining opt-out requires a full page parse
* Not re-usable for generic shadow namespaces

Instead, this adds a __NOGLOBAL__ magic word, which if present on the
central user page, will prevent it from being displayed on remote wikis.
Additionally, this is checked when determining whether a link should be
red or blue.

The old way of opt-out still works, but will be removed in the future
after a deprecation period and time for migration to the new magic word.

Bug: T90849
Change-Id: I29a45bb40b16516f15dfff9a445080949b7fb221