Remove GlobalPreferences integration
At the time, this was a way that people could disable global user page
functionality from showing up, for whatever reason they wanted.

Since it was first implemented, it's been superseded by two other
features. First, the usage of <noinclude></noinclude> to make user
pages appear empty to remote wikis. This behavior was hackish and not
really considered a full solution, so GlobalPreferences support was
kept in. Then we introduced the __NOGLOBAL__ magic word to opt-out
of global user pages. This solution was scalable and non-hackish
and fully supersedes the need for GlobalPreferences integration.

Bug: T188423
Change-Id: Ia204d4c6d9f3fe0389b497fc3b6201bb4c3f4b82