Support new style of table of contents
Vector 2022 skin generates the table of contents from ParserOutput
properties, instead of putting it in the page HTML. Fetch the
necessary data from the API and put it in a ParserOutput. We need to use
TOCData::fromLegacy() since the API only outputs section data in that
format.

To avoid duplicate or missing tables of contents when the local
and remote wikis are using different skins, pass the skin in
the API query as well.

Bug: T327942
Change-Id: If250aeeafd14682a4e6fd2cfb989546db914b180