Track external links to support RealMe
RealMe looks at external links in the ParserOutput to verify the URLs
set in preferences are also on the wiki page. To support this, we now
fetch the list of external links from the remote wiki and add it to the
ParserOutput given to OutputPage, which triggers the RealMe hook.

There is a small risk that a user page has too many external links that
trying to output them sends us over the API result limit, but that's
such an edge case that it's on the user to fix.

== Test plan ==
If you don't have a wiki farm set up locally, you can:
* Hack shouldDisplayGlobalPage() to always return true.
* Point $wgGlobalUserPageDBname at your local wiki and
  $wgGlobalUserPageAPIUrl at Meta-Wiki.

Then:
* Set a URL on your Meta-Wiki userpage in your local preferences as a
  RealMe profile URL.
* View your global user page, see the correct <link rel="me"> element.

Bug: T334805
Change-Id: I553dd5f7f4981fcded9567221bf5d54b40a15a64