Hooks: Reduce overhead of repeated Title->isAlwaysKnown calls
Follows-up I5029dab051e to SpecialBrokenRedirects, which adds
a call to Title->isKnown, which thus calls the onTitleIsAlwaysKnown
hook for each item separately.

* Before: https://performance.wikimedia.org/excimer/profile/3f941003cdb65d67
* After: https://performance.wikimedia.org/excimer/profile/c89ffa65fc355df8
* Comment: Left Heavy, search for "Known".

This probably is okay to leave unbatched since most items will not
be a User page, but even for the items that are not User pages,
we still call various functions before we get to the canBeGlobal()
function and its Title::inNamespace() check for early return.

Improve this by injecting GlobalUserPageManager as dependency for
the hook handler, rather than the static wrapper which re-obtains
service container, and the service object, each time. This makes the
code easier to follow as well by having less indirection.

Idem for the WikiPage subclass (GlobalUserPage).

A potential downside here would be that if most of the hooks or most
of our WikiPage methods do not need GlobalUserPageManager, and if
the constructor for that is expensive, we'd not want to pay that
upfront for the hook handler and/or WikiPage construction. The
constructor for GlobalUserPageManager is cheap, and virtually all
methods and hooks invoke GlobalUserPageManager methods at least once,
so it seems a net-win.

Bug: T90978
Change-Id: I7aa1d676db5df1eb20c74a0c32c7cc5e7664c78e